move in a src folder some of the code
