+++ prec.js
@@ -8,8 +8,8 @@
 // Import some libraries
 const N3            = require('n3');
 
-const RDFGraphBuilder = require("./prec3/graph-builder.js");
-const graphReducer    = require("./prec3/graph-reducer.js");
+const RDFGraphBuilder = require("./src/prec/graph-builder.js");
+const graphReducer    = require("./src/prec/graph-reducer.js");
 
 const { ArgumentParser } = require('argparse');
