+++ src/prec/quad-star-multinested-store.js
